refactor : mock.service코드에 있는 데이터와 계산식 분리작업
- room.ts파일을 만들어 데이터를 따로 저장함
- 유틸폴더를 생성하여 폴더안에 가까운순서에 필요한 함수또는 계산식 분리하여 저장
